# Postman Collection/openapi.json
{"openapi":"3.0.1","info":{"title":"Transaction Report API","version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/transactions/{id}":{"get":{"tags":["transaction-controller"],"operationId":"getTransactionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionDTO"}}}}}},"put":{"tags":["transaction-controller"],"operationId":"updateTransaction","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionDTO"}}}}}},"delete":{"tags":["transaction-controller"],"operationId":"deleteTransaction","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/taxes/{id}":{"get":{"tags":["tax-controller"],"operationId":"getTaxById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaxDTO"}}}}}},"put":{"tags":["tax-controller"],"operationId":"updateTax","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaxDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaxDTO"}}}}}},"delete":{"tags":["tax-controller"],"operationId":"deleteTax","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/products/{id}":{"get":{"tags":["product-controller"],"operationId":"getProductById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"put":{"tags":["product-controller"],"operationId":"updateProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/transactions":{"get":{"tags":["transaction-controller"],"operationId":"getAllTransactions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TransactionDTO"}}}}}}},"post":{"tags":["transaction-controller"],"operationId":"createTransaction","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionDTO"}}}}}}},"/api/taxes":{"get":{"tags":["tax-controller"],"operationId":"getAllTaxes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaxDTO"}}}}}}},"post":{"tags":["tax-controller"],"operationId":"createTax","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaxDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaxDTO"}}}}}}},"/api/products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"/api/auth/signin":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/reports/product-sales":{"get":{"tags":["report-controller"],"operationId":"getProductSalesReport","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/reports/monthly-sales":{"get":{"tags":["report-controller"],"operationId":"getMonthlySalesReport","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number"}}}}}}},"/api/reports/daily-summary":{"get":{"tags":["report-controller"],"operationId":"getDailySummaryReport","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/customers":{"get":{"tags":["customer-controller"],"operationId":"getAllCustomers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"UserDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","SUSPENDED"]},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"createdBy":{"type":"string","format":"uuid"},"updatedBy":{"type":"string","format":"uuid"}},"example":{"username":"newuser","email":"newuser@example.com","firstName":"New","lastName":"User","phoneNumber":"081234567899","status":"ACTIVE","roles":["STAFF"]}},"TransactionDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"transactionNumber":{"type":"string"},"customerId":{"type":"string","format":"uuid"},"netAmountPaid":{"type":"number"},"totalTaxPaid":{"type":"number"},"totalAmountPaid":{"type":"number"},"transactionTime":{"type":"string","format":"date-time"},"paymentStatus":{"type":"string","enum":["PAID","UNPAID","CANCELLED","REFUNDED"]},"paymentMethod":{"type":"string","enum":["CASH","CREDIT_CARD","DEBIT_CARD","BANK_TRANSFER","DIGITAL_WALLET"]},"items":{"type":"array","items":{"$ref":"#/components/schemas/TransactionItemDTO"}},"createdBy":{"type":"string","format":"uuid"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"notes":{"type":"string"}},"example":{"transactionNumber":"TRX-20240721-001","customerId":"0f75f8f9-fd4a-4b80-9ed6-ecf3c0863abf","netAmountPaid":100.0,"totalTaxPaid":10.0,"totalAmountPaid":110.0,"transactionTime":"2024-07-21T10:00:00","paymentStatus":"PAID","paymentMethod":"CASH","items":[{"productId":"8f6ef676-f9e5-402f-9338-55a8b406eae6","quantity":1,"unitPrice":100.0,"totalPrice":100.0,"taxAmount":10.0}],"notes":"Sample transaction"}},"TransactionItemDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"productId":{"type":"string","format":"uuid"},"quantity":{"type":"integer","format":"int32"},"unitPrice":{"type":"number"},"totalPrice":{"type":"number"},"taxAmount":{"type":"number"}},"example":{"productId":"8f6ef676-f9e5-402f-9338-55a8b406eae6","quantity":1,"unitPrice":100.0,"totalPrice":100.0,"taxAmount":10.0}},"TaxDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"rate":{"type":"number"},"type":{"type":"string","enum":["PERCENTAGE","FIXED_AMOUNT"]},"description":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","INACTIVE"]}},"example":{"name":"VAT 10%","rate":10.0,"type":"PERCENTAGE","description":"Value Added Tax 10%","status":"ACTIVE"}},"ProductDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number"},"status":{"type":"string","enum":["ACTIVE","INACTIVE","DISCONTINUED"]},"category":{"type":"string"},"sku":{"type":"string"},"taxIds":{"uniqueItems":true,"type":"array","items":{"type":"string","format":"uuid"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"example":{"name":"Laptop XYZ","description":"Powerful laptop for everyday use","price":1200.0,"status":"ACTIVE","category":"Electronics","sku":"LPT-XYZ-001","taxIds":[]}},"LoginRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"example":{"username":"admin","password":"adminpass"}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}